+ Refactor codes
Resolves #12 and #13
